add support for tinyblob
datatypes blob and tinyblob needs to be supported. currently the row of data will only be logged and this fix adds support for blobdata.
Base64 encoding ensures that the binary data is preserved without any loss during serialization and deserialization processes, making it a reliable method for transmitting binary data in textual formats like JSON.